# openapi.yaml
...
openapi: 3.1.0
info:
  title: Kryton API
  version: 1.1.0
  description: |
    Kryton Windows virtualization control plane. All machine operations are project-scoped.
    Authenticate with `Authorization: Bearer <token>` (apikey mode) or trusted proxy headers.
    OpenAPI is served at `/openapi.yaml` and discovery at `GET /api/v1`.
  contact:
    name: Zyvor
    url: https://zyvor.dev
  license:
    name: Apache-2.0
servers:
  - url: /
    description: Relative to Kryton base URL
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/storage/config:
    get:
      tags: [storage]
      summary: Active default StorageClass
      responses:
        '200': {description: Storage config}
    put:
      tags: [storage]
      summary: Set default StorageClass for new VM disks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storageClass: {type: string, example: rook-ceph-block}
      responses:
        '200': {description: Updated}
        '400': {description: Invalid class}
  /api/v1/storage/setup:
    get:
      tags: [storage]
      summary: Storage install job status and logs
      responses:
        '200': {description: Setup status}
    post:
      tags: [storage]
      summary: Install Longhorn or Rook Ceph
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageSetup'
      responses:
        '202': {description: Setup started}
        '400': {description: Invalid request}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from KRYTON_API_KEYS_FILE (hash with krytonctl hash-token)
...
  schemas:
...
    StorageSetup:
      type: object
      required: [backend]
      properties:
        backend: {type: string, enum: [longhorn, rook-ceph]}
        rookMode: {type: string, enum: [lab, devices, pool-only, device]}
        device: {type: string, example: /dev/sdb1}
        wipeDevice: {type: boolean}
        setDefault: {type: boolean, default: true}
...